[#159] SwiftData 폴더링 구조 변경 및 분리
