+++ MC2-OneShot/Data/Enum/NotiCycle.swift
