+++ MC2-OneShot/Data/Path/CameraPath.swift
