+++ MC2-OneShot/Data/Path/PathModel.swift
