+++ MC2-OneShot/Data/SwiftData/Media.swift
@@ -1,0 +1,22 @@
+//
+//  Media.swift
+//  MC2-OneShot
+//
+//  Created by 김민준 on 6/1/24.
+//
+
+import Foundation
+import SwiftData
+
+// MARK: - Media
+
+@Model
+class Media: Identifiable {
+    var fileData: Data
+    var captureDate: Date
+    
+    init(fileData: Data, captureDate: Date) {
+        self.fileData = fileData
+        self.captureDate = captureDate
+    }
+}
